refactor architecture
