--- domain/entity/Client.go
@@ -1,8 +1,0 @@
-package entity
-
-type Client struct {
-	ClientId int
-	FirstName string
-	LastName string
-	Accounts []Account
-}
