--- domain/repo/StorageRepo.go
@@ -1,5 +1,0 @@
-package domain
-
-type StorageRepo interface{
-	Read() [][]string 
-}
