+++ internal/domain/Client.go
@@ -1,0 +1,8 @@
+package domain
+
+type Client struct {
+	ClientId  int
+	FirstName string
+	LastName  string
+	Accounts  []Account
+}
