+++ internal/domain/Currency.go
@@ -1,4 +1,4 @@
-package entity
+package domain
 
 type Currency int
