+++ internal/service/account.go
@@ -1,0 +1,5 @@
+package service
+
+type AccountStorage interface{
+	Read() [][]string 
+}
